Added fucntionaility for save and delete buttons on EditItemDialog

+++ app/src/main/java/com/srg/mealmate/GroceryItem.java
@@ -57,7 +57,7 @@ public class GroceryItem implements Serializable {
         isChecked = tf;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
